requirements.txt
#Flask
#Flask-SQLAlchemy
#Flask-Session
#Flask-Caching
#Flask-Migrate
#Flask-Script
#SQLAlchemy
#SQLAlchemy-Utils
flask-caching==1.8.0      # via -r requirements.in
flask-marshmallow==0.10.1  # via -r requirements.in
flask-migrate==2.5.3      # via -r requirements.in
flask-restx==0.2.0        # via -r requirements.in
flask-script==2.0.6       # via -r requirements.in
flask-sqlalchemy==2.4.3   # via -r requirements.in, flask-migrate
flask==1.1.2              
sqlalchemy-utils==0.36.6  # via -r requirements.in
sqlalchemy==1.3.17      
passlib==1.7.1
bcrypt==3.1.4
six==1.11.0
itsdangerous==1.1.0
requests>=2.20.0
PyMySQL==0.9.2
gunicorn==20.0.4
normality==0.6.1  # Pin normality version for dataset
dataset==1.1.0
mistune==0.8.4
netaddr==0.7.19
redis==2.10.6
datafreeze==0.1.0
gevent-websocket
python-dotenv==0.9.1
flask-restplus
pathlib2==2.3.2
marshmallow-sqlalchemy==0.17.0  # via -r requirements.in
marshmallow==2.20.2       # via -r requirements.in, flask-marshmallow, marshmallow-sqlalchemy
Flask-SocketIO==3.0.2
boto3==1.9.39
werkzeug==1.0.1
cmarkgfm==0.4.2
gevent==20.9.0            # via -r requirements.in
greenlet==0.4.17          # via gevent
discord.py
